requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0/de_core_news_sm-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
fastapi==0.63.0
filelock==3.0.12
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0-py3-none-any.whl
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
pathspec==0.8.1
pathy==0.5.0
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.0
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sacremoses==0.0.45
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-alignments==0.8.3
spacy-legacy==3.0.4
spacy-lookups-data==1.0.0
spacy-transformers==1.0.2
srsly==2.4.1
starlette==0.13.6
thinc==8.0.3
tokenizers==0.10.2
toml==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.1
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
wasabi==0.8.2
